# Faster `differences_to_samples`