This function is responsible for 55% of execution time after #173; optimizations here have a large effect on performance.

This change leverages instruction-level parallelism to speed it up.

Benchmarks before (current master):
```
test read_single_image_non_parallel_zips_rgba        ... bench: 196,783,469 ns/iter (+/- 172,721)
test read_single_image_rle_all_channels              ... bench:  41,137,521 ns/iter (+/- 3,633,097)
test read_single_image_rle_non_parallel_all_channels ... bench:  98,932,909 ns/iter (+/- 2,856,674)
test read_single_image_uncompressed_from_buffer_rgba ... bench:  33,522,367 ns/iter (+/- 70,401)
test read_single_image_uncompressed_rgba             ... bench:  37,435,859 ns/iter (+/- 2,202,515)
test read_single_image_zips_rgba                     ... bench:  71,419,820 ns/iter (+/- 4,860,501)
```

Benchmarks after:

```
test read_single_image_non_parallel_zips_rgba        ... bench: 183,337,110 ns/iter (+/- 4,920,259)
test read_single_image_rle_all_channels              ... bench:  38,442,440 ns/iter (+/- 4,608,694)
test read_single_image_rle_non_parallel_all_channels ... bench:  81,506,752 ns/iter (+/- 796,542)
test read_single_image_uncompressed_from_buffer_rgba ... bench:  31,839,775 ns/iter (+/- 81,663)
test read_single_image_uncompressed_rgba             ... bench:  34,882,260 ns/iter (+/- 86,133)
test read_single_image_zips_rgba                     ... bench:  69,871,756 ns/iter (+/- 6,176,810)
```

The effect is more pronounced when combined with #173  (as in, this change makes up a larger percentage of execution time).

Tests pass, and they did fail on earlier versions where I messed them up slightly.

If this approach is OK with you, I can also try to optimize `samples_to_differences` in a similar fashion, although that's more difficult because it iterates in reverse and references future values instead of past values.