@Shnatsel 2022-12-29
I've also optimized compression in a similar way, for modest but noticeable performance improvement across the board.

Before:
```
test write_nonparallel_zip1_to_buffered      ... bench: 481,441,257 ns/iter (+/- 22,410,241)
test write_parallel_any_channels_to_buffered ... bench:  67,985,884 ns/iter (+/- 18,264,908)
test write_parallel_zip16_to_buffered        ... bench: 171,812,583 ns/iter (+/- 8,102,075)
test write_parallel_zip1_to_buffered         ... bench: 142,734,713 ns/iter (+/- 7,649,419)
test write_uncompressed_to_buffered          ... bench:  25,297,916 ns/iter (+/- 187,416)
```

After:
```
test write_nonparallel_zip1_to_buffered      ... bench: 468,520,748 ns/iter (+/- 7,942,493)
test write_parallel_any_channels_to_buffered ... bench:  62,532,362 ns/iter (+/- 7,417,551)
test write_parallel_zip16_to_buffered        ... bench: 165,498,070 ns/iter (+/- 5,988,365)
test write_parallel_zip1_to_buffered         ... bench: 139,177,276 ns/iter (+/- 10,863,506)
test write_uncompressed_to_buffered          ... bench:  24,443,239 ns/iter (+/- 149,205)
```

If there were a single-threaded RLE write benchmark, it would show a big improvement.