@Shnatsel 2022-12-30
I've optimized `samples_to_differences` further. It's now 5x faster than the naive version, or 10x faster with `-C target-cpu=native`.